[shared] re-organize types into sub-modules
