+++ apps/src/bin/anoma-client/cli.rs
@@ -5,9 +5,10 @@ use anoma::client::tx;
 use anoma::proto::services::rpc_service_client::RpcServiceClient;
 use anoma::proto::{self, services, RpcMessage};
 use anoma::{cli, wallet};
+use anoma_shared::types::address::Address;
 use anoma_shared::types::intent::Intent;
 use anoma_shared::types::key::ed25519::Signed;
-use anoma_shared::types::{token, Address};
+use anoma_shared::types::token;
 use borsh::BorshSerialize;
 use color_eyre::eyre::Result;
 use eyre::Context;
