+++ apps/src/lib/client/tx.rs
@@ -1,7 +1,9 @@
 use std::str::FromStr;
 
+use anoma_shared::types::address::Address;
 use anoma_shared::types::key::ed25519::Keypair;
-use anoma_shared::types::{token, Address, UpdateVp};
+use anoma_shared::types::token;
+use anoma_shared::types::transaction::UpdateVp;
 use borsh::BorshSerialize;
 use tendermint_rpc::{Client, HttpClient};
