+++ shared/src/types/key/ed25519.rs
@@ -12,7 +12,8 @@ use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use thiserror::Error;
 
-use crate::types::{address, Address, DbKeySeg, Key, KeySeg};
+use crate::types::address::{self, Address};
+use crate::types::storage::{DbKeySeg, Key, KeySeg};
 
 const SIGNATURE_LEN: usize = ed25519_dalek::SIGNATURE_LENGTH;
