+++ shared/src/types/token.rs
@@ -3,8 +3,9 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 
+use crate::types::address::Address;
 use crate::types::key::ed25519::{Keypair, SignedTxData};
-use crate::types::{Address, DbKeySeg, Key, KeySeg};
+use crate::types::storage::{DbKeySeg, Key, KeySeg};
 
 /// Amount in micro units. For different granularity another representation
 /// might be more appropriate.
