+++ shared/src/vm/host_env.rs
@@ -10,9 +10,10 @@ use crate::gossip::mm::MmHost;
 use crate::ledger::gas::{BlockGasMeter, VpGasMeter};
 use crate::ledger::storage::write_log::{self, WriteLog};
 use crate::ledger::storage::{self, Storage, StorageHasher};
+use crate::types::address::Address;
 use crate::types::internal::HostEnvResult;
 use crate::types::key::ed25519::{verify_signature_raw, PublicKey, Signature};
-use crate::types::{Address, Key};
+use crate::types::storage::Key;
 use crate::vm::memory::VmMemory;
 use crate::vm::prefix_iter::{PrefixIteratorId, PrefixIterators};
 use crate::vm::types::KeyVal;
