+++ shared/src/vm/wasm/runner.rs
@@ -16,8 +16,9 @@ use crate::gossip::mm::MmHost;
 use crate::ledger::gas::{BlockGasMeter, VpGasMeter};
 use crate::ledger::storage::write_log::WriteLog;
 use crate::ledger::storage::{self, Storage, StorageHasher};
+use crate::types::address::Address;
 use crate::types::internal::HostEnvResult;
-use crate::types::{Address, Key};
+use crate::types::storage::Key;
 use crate::vm::host_env::VpEvalRunner;
 use crate::vm::prefix_iter::PrefixIterators;
 use crate::vm::types::{TxInput, VpInput};
