+++ tests/src/vm_host_env/tx.rs
@@ -4,7 +4,8 @@ use anoma_shared::ledger::gas::BlockGasMeter;
 use anoma_shared::ledger::storage::mockdb::MockDB;
 use anoma_shared::ledger::storage::testing::TestStorage;
 use anoma_shared::ledger::storage::write_log::WriteLog;
-use anoma_shared::types::{Address, Key};
+use anoma_shared::types::address::Address;
+use anoma_shared::types::storage::Key;
 use anoma_shared::vm;
 use anoma_shared::vm::prefix_iter::PrefixIterators;
