+++ vm_env/src/imports.rs
@@ -5,9 +5,10 @@ pub mod tx {
     use std::marker::PhantomData;
     pub use std::mem::size_of;
 
+    use anoma_shared::types::address::Address;
     use anoma_shared::types::internal::HostEnvResult;
-    use anoma_shared::types::{
-        Address, BlockHash, BlockHeight, BLOCK_HASH_LENGTH, CHAIN_ID_LENGTH,
+    use anoma_shared::types::storage::{
+        BlockHash, BlockHeight, BLOCK_HASH_LENGTH, CHAIN_ID_LENGTH,
     };
     use anoma_shared::vm::types::KeyVal;
     pub use borsh::{BorshDeserialize, BorshSerialize};
@@ -244,7 +245,7 @@ pub mod vp {
 
     use anoma_shared::types::internal::HostEnvResult;
     use anoma_shared::types::key::ed25519::{PublicKey, Signature};
-    use anoma_shared::types::{
+    use anoma_shared::types::storage::{
         BlockHash, BlockHeight, BLOCK_HASH_LENGTH, CHAIN_ID_LENGTH,
     };
     use anoma_shared::vm::types::KeyVal;
