+++ wasm/vps/vp_template/src/lib.rs
@@ -6,7 +6,7 @@ use anoma_vm_env::vp_prelude::*;
 fn validate_tx(
     tx_data: Vec<u8>,
     addr: Address,
-    keys_changed: Vec<Key>,
+    keys_changed: Vec<storage::Key>,
     verifiers: HashSet<Address>,
 ) -> bool {
     log_string(format!(
