+++ wasm/vps/vp_user/src/lib.rs
@@ -10,8 +10,8 @@ enum KeyType<'a> {
     Unknown,
 }
 
-impl<'a> From<&'a Key> for KeyType<'a> {
-    fn from(key: &'a Key) -> KeyType<'a> {
+impl<'a> From<&'a storage::Key> for KeyType<'a> {
+    fn from(key: &'a storage::Key) -> KeyType<'a> {
         if let Some(address) = token::is_any_token_balance_key(key) {
             Self::Token(address)
         } else if let Some(address) = intent::is_invalid_intent_key(key) {
@@ -26,7 +26,7 @@ impl<'a> From<&'a Key> for KeyType<'a> {
 fn validate_tx(
     tx_data: Vec<u8>,
     addr: Address,
-    keys_changed: Vec<Key>,
+    keys_changed: Vec<storage::Key>,
     verifiers: HashSet<Address>,
 ) -> bool {
     log_string(format!(
@@ -199,7 +199,7 @@ mod tests {
 
         let tx_data: Vec<u8> = vec![];
         let addr: Address = env.addr.clone();
-        let keys_changed: Vec<Key> = vec![];
+        let keys_changed: Vec<storage::Key> = vec![];
         let verifiers: HashSet<Address> = HashSet::default();
 
         let valid = validate_tx(tx_data, addr, keys_changed, verifiers);
